+++ HISTORY.rst
@@ -18,6 +18,8 @@ End-User Summary
 - Included MutationTaster as additional pathogenicity score.
 - Included UMD-Predictor as additional pathogenicity score.
 - Project-wide filter now applicable when the project contains cases with no small variants (e.g. completely empty or only SVs).
+- Ignoring option `remove if in dbSNP` when `ClinVar membership required` is activated as every ClinVar entry has a dbSNP id.
+- Fixed indices on `SmallVariantFlags` and `SmallVariantComment` and introduced indices for `ExacConstraints` and `GnomadConstraints` that sped up large queries significantly.
 
 Full Change List
 ================
@@ -34,6 +36,8 @@ Full Change List
 - Included MutationTaster as additional pathogenicity score.
 - Included UMD-Predictor as additional pathogenicity score.
 - Project-wide filter now applicable when the project contains cases with no small variants (e.g. completely empty or only SVs).
+- Ignoring option `remove if in dbSNP` when `ClinVar membership required` is activated as every ClinVar entry has a dbSNP id.
+- Fixed indices on `SmallVariantFlags` and `SmallVariantComment` and introduced indices for `ExacConstraints` and `GnomadConstraints` that sped up large queries significantly.
 
 -------
 v0.17.4